+++ src/backend/microservices/dataset_loading/.env
@@ -1,0 +1,4 @@
+DB_HOST=178.211.133.63
+DB_USER=wsozgsko_datadive_backend
+DB_PASS=2Y[-*^${*n{6
+DB_NAME=wsozgsko_datadive_backend
