+++ src/backend/microservices/dataset_loading/requirements.txt
@@ -7,4 +7,3 @@ openpyxl
 uuid
 mysql.connector
 python-dotenv
-logging
